+++ RailTrak/RailTrak/RailTrak.py
@@ -1,1 +1,16 @@
+from pywebio.input import *
+from pywebio.output import *
+from pywebio import start_server
+from UserDBM import UserDBM
+from TrackGUI import *
+from history_log import showHistoryLog
+from lloginPage import loginPage
 
+# Index / homepage of the RailTrac website
+def index():
+    put_markdown(r""" # RailTrac Homepage
+    """)
+    put_link('Returning User Login', app='loginPage')
+
+# Start the RailTrac application
+start_server([index, loginPage, Selection, showHistoryLog], port=80, debug=True, remote_access=True)
